# src/main/openapi/api-docs.yaml
openapi: 3.1.0
info:
  title: SmartDoc Document Management API
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  description: API for managing person-related documents stored in a file system, and for serving SmartDoc UI plugin resources. Designed to be accessed via the OH-API proxy in integrated mode.
  version: 1.1.0
servers:
  - url: http://localhost:8080/api
security:
  - Bearer: []
...
paths:
  /documents:
    get:
      tags:
        - Documents
      summary: Retrieve a paginated and filtered list of documents.
      operationId: findDocuments
      parameters:
        - $ref: '#/components/parameters/PersonId'
        - $ref: '#/components/parameters/DocumentType'
        - $ref: '#/components/parameters/FilterFromDate'
        - $ref: '#/components/parameters/FilterToDate'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Successful retrieval of document list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDocument'
        '400':
          description: Invalid query parameters provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags:
        - Documents
      summary: Upload a new document file for a person.
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadDocumentRequest'
      responses:
        '201':
          description: Document successfully uploaded and stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid request, file format/size, or missing required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error (file system write error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /documents/{id}:
    get:
      tags:
        - Documents
      summary: Retrieve document metadata.
      operationId: findDocumentById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Document metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: Document not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      tags:
        - Documents
      summary: Update document metadata or replace the file.
      operationId: updateDocument
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadDocumentRequest'
      responses:
        '200':
          description: Document successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid request, file format/size, or missing required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error (file system write error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags:
        - Documents
      summary: Permanently delete a document file from the file system.
      operationId: deleteDocument
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Document successfully deleted.
        '404':
          description: Document not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: File system deletion error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /documents/{id}/download:
    get:
      tags:
        - Documents
      summary: Download a document file
      operationId: downloadDocument
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Attachment'
      responses:
        '200':
          description: Document file download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /persons:
    get:
      tags:
        - Persons
      summary: List persons with pagination (used for stand-alone mode or autocomplete lookup).
      operationId: findPersons
      parameters:
        - $ref: '#/components/parameters/PersonNamePartial'
        - $ref: '#/components/parameters/IncludeInactive'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Paginated list of persons.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPerson'
        '400':
          description: Bad request (invalid parameters).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags:
        - Persons
      summary: Create a new person.
      operationId: createPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonRequest'
      responses:
        '201':
          description: Person created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: Bad request (invalid data).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden (insufficient permissions).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /persons/{id}:
    get:
      tags:
        - Persons
      summary: Retrieve a person by ID.
      operationId: findPersonById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Person found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '403':
          description: Forbidden (insufficient permissions).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Person not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      tags:
        - Persons
      summary: Update a person.
      operationId: updatePerson
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePersonRequest'
      responses:
        '200':
          description: Person updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: Bad request (invalid data).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden (insufficient permissions).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Person not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      tags:
        - Persons
      summary: Partially update a person.
      operationId: patchPerson
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchPersonRequest'
      responses:
        '200':
          description: Person updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: Bad request (invalid data).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden (insufficient permissions).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Person not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags:
        - Persons
      summary: Delete a person.
      operationId: deletePerson
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Person deleted.
        '403':
          description: Forbidden (insufficient permissions).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Person not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /persons/{id}/documents:
    get:
      tags:
        - Persons
      summary: Retrieve a paginated and filtered list of documents for a specific person.
      operationId: findPersonDocuments
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/DocumentType'
        - $ref: '#/components/parameters/FilterFromDate'
        - $ref: '#/components/parameters/FilterToDate'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Successful retrieval of document list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDocument'
        '400':
          description: Invalid query parameters provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Person not found (No directory exists for the personId).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /persons/{id}/restore:
    post:
      tags:
        - Persons
      summary: Restore a deleted person.
      operationId: restorePerson
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Person restored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: Person not deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden (insufficient permissions).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Person not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /persons/{id}/activate:
    post:
      tags:
        - Persons
      summary: Activate an inactive person.
      operationId: activatePerson
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Person activated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: Person not deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden (insufficient permissions).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Person not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /persons/{id}/deactivate:
    post:
      tags:
        - Persons
      summary: Deactivate an active person.
      operationId: deactivatePerson
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Person deactivated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: Person not deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden (insufficient permissions).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Person not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    Basic:
      type: http
      scheme: basic
      description: Basic authentication
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token
  parameters:
    PersonId:
      name: personId
      in: query
      schema:
        type: string
        pattern: ^\d{1,6}$
      description: Numeric identifier of the person (up to 6 digits).
    DocumentType:
      name: type
      in: query
      required: true
      schema:
        type: string
        enum:
          - DOC
          - MED
          - REF
      description: Document category type (DOC, MED, REF). Used to target a specific subfolder.
    FilterFromDate:
      name: fromDate
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: Filter documents starting from this clinical date (YYYY-MM-DD), parsed from the filename.
    FilterToDate:
      name: toDate
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: Filter documents up to this clinical date (YYYY-MM-DD), parsed from the filename.
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
      description: Page number for pagination.
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 1
      description: Number of items per page for pagination.
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Numeric identifier of the resource.
    Attachment:
      name: attachment
      in: query
      schema:
        type: boolean
        default: true
      description: If true, sets Content-Disposition to attachment for download
    PersonNamePartial:
      name: name
      in: query
      required: false
      schema:
        type: string
      description: Partial name for filtering (used when proxied by OH-API for OH data lookup).
    IncludeInactive:
      name: includeInactive
      in: query
      description: Include inactive entities in the results.
      required: false
      schema:
        type: boolean
        default: false
  schemas:
    Version:
      type: integer
      format: int64
      description: The lock field for the object.
    Base:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the object.
        createdBy:
          type: string
          description: The user who created the object.
        createdDate:
          type: string
          format: date-time
          description: The date and time when the object was created.
        lastModifiedBy:
          type: string
          description: The user who last updated the object.
        lastModifiedDate:
          type: string
          format: date-time
          description: The date and time when the object was last updated.
        version:
          $ref: '#/components/schemas/Version'
      required:
        - id
        - createdDate
        - lastModifiedDate
    DocumentStatus:
      type: string
      enum:
        - active
        - archived
        - deleted
      description: Document lifecycle status values.
    Document:
      allOf:
        - $ref: '#/components/schemas/Base'
        - type: object
          description: Detailed information about a single document.
          properties:
            fileName:
              type: string
              description: The full file name in the storage.
            path:
              type: string
              description: The path of the document in the system.
            personId:
              type: string
            type:
              type: string
              description: The document type code.
            date:
              type: string
              format: date-time
              description: The clinical date extracted from the file name.
            description:
              type: string
              description: Human-readable description of the document content.
            fileSize:
              type: integer
              description: File size in bytes.
            mimeType:
              type: string
              description: MIME type of the file.
            status:
              $ref: '#/components/schemas/DocumentStatus'
            uploadDate:
              type: string
              format: date-time
              description: Timestamp of document upload.
    PageInfo:
      type: object
      description: Pagination information.
      required:
        - page
        - size
        - totalElements
        - totalPages
      properties:
        page:
          type: integer
          description: Current page number (0-based).
          example: 0
        size:
          type: integer
          description: Number of items per page.
          example: 10
        totalElements:
          type: integer
          description: Total number of elements across all pages.
          example: 57
        totalPages:
          type: integer
          description: Total number of pages.
          example: 6
    PaginatedDocument:
      type: object
      description: Structure for a paginated list of documents.
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Document'
          description: List of documents for the current page.
        metadata:
          $ref: '#/components/schemas/PageInfo'
    Problem:
      type: object
      description: Standardized error format following RFC 7807 (Problem Details for HTTP APIs).
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          example: https://my.domain.xd/problems/invalid-parameter
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: Validation Error
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          example: Required field 'personId' is missing or malformed.
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of the problem.
        traceId:
          type: string
          format: uuid
          description: A unique identifier for the error occurrence, useful for tracing.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    BinaryFile:
      type: string
      format: binary
      description: Represents raw binary file content (JPEG, JavaScript bundle).
    DocumentMetadata:
      type: object
      required:
        - personId
        - type
      properties:
        personId:
          type: string
          format: uuid
          description: Numeric identifier of the person.
        type:
          type: string
          format: uuid
          description: Document category type id.
        date:
          type: string
          format: date
          description: The clinical date associated with the document (YYYY-MM-DD). Used for the file naming timestamp.
        description:
          type: string
          description: Optional description to be included in the filename.
      description: Document metadata associated with an uploaded file.
    UploadDocumentRequest:
      type: object
      required:
        - document
        - metadata
      properties:
        document:
          allOf:
            - $ref: '#/components/schemas/BinaryFile'
          description: The file to upload (JPG).
        metadata:
          $ref: '#/components/schemas/DocumentMetadata'
      description: Fields required for document upload via multipart/form-data.
    Gender:
      type: string
      enum:
        - male
        - female
      description: Administrative gender values.
    Status:
      type: string
      enum:
        - active
        - inactive
        - deleted
      description: Entity status enumeration supporting soft deletion
      default: active
    Person:
      allOf:
        - $ref: '#/components/schemas/Base'
        - type: object
          description: Person identification data.
          properties:
            name:
              type: string
              description: Person's name (available when integrated with OH).
            pid:
              type: string
              description: The patient external identifier.
            email:
              type: string
              format: email
              description: Contact email.
            phoneNumber:
              type: string
              description: Primary phone number.
            gender:
              $ref: '#/components/schemas/Gender'
            status:
              $ref: '#/components/schemas/Status'
          required:
            - name
            - pid
            - status
    PaginatedPerson:
      type: object
      description: Structure for a paginated list of persons.
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Person'
          description: List of persons for the current page.
        metadata:
          $ref: '#/components/schemas/PageInfo'
    CreatePersonRequest:
      type: object
      properties:
        name:
          type: string
          description: Person's name (available when integrated with OH).
        pid:
          type: string
          description: The patient external identifier.
        email:
          type: string
          format: email
          description: Contact email.
        phoneNumber:
          type: string
          description: Primary phone number.
        gender:
          $ref: '#/components/schemas/Gender'
      required:
        - name
        - pid
    UpdatePersonRequest:
      type: object
      properties:
        version:
          $ref: '#/components/schemas/Version'
        name:
          type: string
          description: Person's name (available when integrated with OH).
        pid:
          type: string
          description: The patient external identifier.
        email:
          type: string
          format: email
          description: Contact email.
        phoneNumber:
          type: string
          description: Primary phone number.
        gender:
          $ref: '#/components/schemas/Gender'
      required:
        - version
        - name
        - pid
    PatchPersonRequest:
      type: object
      properties:
        version:
          $ref: '#/components/schemas/Version'
        name:
          type: string
          description: Person's name (available when integrated with OH).
        pid:
          type: string
          description: The patient external identifier.
        email:
          type: string
          format: email
          description: Contact email.
        phoneNumber:
          type: string
          description: Primary phone number.
        gender:
          $ref: '#/components/schemas/Gender'
...